refactor: add dialog flow hooks
